Extract StringConverter for paymentAccountComboBox to GUIUtils. Handle multicurrency accounts.
